Stop using SQLAlchemy in the insert script + add documentation
So the insert script has been modified now to NOT use SQLAlchemy,
the code directly generates insert statements and executes them.
This was done because the previous method was extremely slow and
we had to insert thousands of tweets per second if the Pub/Sub queue
was to be processed in a timely fashion.

The script was also documented with comments and some cleanup.