Constructing a machine learning model to categorize tweets as containing a location or not
The Jupyter Notebook gathers the data into a pandas dataframe and performs grid-search
followed by fitting with optimal parameters for a random forests model, to categorize
tweets as containing a location or not. This model is then applied to all tweets
previously flagged as containing a location, and the results are written to a geojson
file that can be used by the Mapbox JS API.

The notebook also generates feature importance and confusion matrix figures.